requirements.txt
Django>=4.2
pandas>=1.5
scikit-learn>=1.2
djangorestframework
django-environ>=0.11
dj-database-url>=2.0
django-cors-headers>=4.0
django-widget-tweaks>=1.4
django-extensions>=3.2
paypalrestsdk>=1.13
requests>=2.28 